integration_team6/requirements.txt
flask
requests
mysql-connector-python
werkzeug
